+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Updated megacmd to `v2.5.2b3`
 - Updated base image to `alpine:3.23`
 - Updated NiceGUI to 3.10
+- Replaced tmux with Python-based terminal session and new NiceGUI xterm UI
 
 ## 0.1.10
 - Went back to NiceGUI `2.20.0` because of refresh bug